+++ sdk/src/lib.rs
@@ -4,10 +4,10 @@ pub mod fee_calculator;
 pub mod genesis_block;
 pub mod hash;
 pub mod instruction;
+pub mod instruction_processor_utils;
 pub mod loader_instruction;
 pub mod message;
 pub mod native_loader;
-pub mod native_program;
 pub mod packet;
 pub mod pubkey;
 pub mod rpc_port;
